app/pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.1"
description = "METNO METCAP API"
authors = ["METNO"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "~1.19.5"
uvicorn = "~0.11.3"
fastapi = "~0.75.0"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "~3.0.1"
tenacity = "~8.0.1"
httpx = "~0.18.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
requests = "^2.26.0"
pyyaml = "^5.3.1"
shapely = "^1.8"
pytest = "^6.2.5"
prometheus-client = "^0.14.1"
couchdb = "^1.2"
pyproj = "^3.3.1"
#
# modules for import_cap.py
#
tqdm  = "^4.64.0"
xmltodict = "^0.13.0"
lxml = "^4.9.1"
geojson-rewind = "^1.0.3"
#
# modules for RSS feeds
#
feedparser = "^6.0.10"

[tool.poetry.dev-dependencies]
